reorder processes, and adding prompt for secure boot function
basically adding a prompt for the user to ask whether or not setup mode is enabled in the bios for secure boot. if not, continues the installation without attempting to forcefully enable secure boot, since not everybody would want it.

also added a --needed flag for the pacman command after the secure boot prompt to download only what hasn't already have been downloaded.