reorder processes, remove secure boot prompt, and added new packages.
the whole shtick of the script is that it's fully automated, so it doesn't make sense for the script to ask for a prompt in the middle of the script while the user is off doing something else.

added some packages to account for archinstall, and also reordered some instructions around.