add "vscodium" to AUR packages to download.
also added the fonts to repo to make downloading easier, and remove the need to properly rename them.

going to start writing the script as if it's going to be run in root, because i want to work towards being able to do everything unhindered, as some commands are ignored despite being run in sudo (e.g. writing files in a system directory).